chore(labels): identify main harness container
Add a new label to identify what container is the entrypoint for debugging
operations in a harness.
