+++ internal/harnesses/container/container.go
@@ -116,6 +116,7 @@ func New(name string, cli *provider.DockerClient, cfg Config) types.Harness {
 				MemoryRequest: resource.MustParse("250Mi"),
 			},
 			Privileged: cfg.Privileged,
+			Labels:     provider.MainHarnessLabel(),
 		},
 		Mounts:         mounts,
 		ManagedVolumes: managedVolumes,
