+++ internal/harnesses/docker/docker.go
@@ -116,6 +116,7 @@ func New(id string, cli *provider.DockerClient, opts ...Option) (types.Harness, 
 				},
 			},
 			Resources: resources,
+			Labels:    provider.MainHarnessLabel(),
 		},
 		Mounts:         mounts,
 		ManagedVolumes: managedVolumes,
