+++ internal/provider/harness_k3s_resource.go
@@ -245,8 +245,8 @@ func (r *HarnessK3sResource) Create(ctx context.Context, req resource.CreateRequ
 	id := data.Id.ValueString()
 	configVolumeName := id + "-config"
 
-	_, err = r.store.cli.VolumeCreate(ctx, volume.CreateOptions{
-		Labels: provider.DefaultLabels,
+	_, err := r.store.cli.VolumeCreate(ctx, volume.CreateOptions{
+		Labels: provider.DefaultLabels(),
 		Name:   configVolumeName,
 	})
 	if err != nil {
